refactor: extract PermissionRecoveryStatusPresenter.swift from PermissionRecoveryTypes.swift
Splits the `PermissionRecoveryStatusPresenter` @MainActor class out of
PermissionRecoveryTypes.swift so that file holds only the
`PermissionRecoveryRefreshOutcome` + `PermissionSettingsOpenResult`
outcome enums.

Byte-identical move of lines 13-28. Both files import Foundation.

PermissionRecoveryTypes.swift: 28 → 11 lines. Tests: 667.
